[build] Preserve normalized workload revisions
Carry the normalized Version through revision selection so single-component feed revisions remain supported, and add CI regression coverage.
