+++ build-tools/automation/yaml-templates/stage-xamarin-android-tools-tests.yaml
@@ -91,6 +91,7 @@ stages:
         command: test
         projects: |
           $(System.DefaultWorkingDirectory)/tests/Microsoft.Android.Build.BaseTasks-Tests/Microsoft.Android.Build.BaseTasks-Tests.csproj
+          $(System.DefaultWorkingDirectory)/tests/WorkloadDependencies-Tests/WorkloadDependencies-Tests.csproj
           $(System.DefaultWorkingDirectory)/tests/Xamarin.Android.Tools.AndroidSdk-Tests/Xamarin.Android.Tools.AndroidSdk-Tests.csproj
         arguments: -c Debug $(XatBuildProperties)
 
@@ -143,6 +144,7 @@ stages:
         command: test
         projects: |
           $(System.DefaultWorkingDirectory)/tests/Microsoft.Android.Build.BaseTasks-Tests/Microsoft.Android.Build.BaseTasks-Tests.csproj
+          $(System.DefaultWorkingDirectory)/tests/WorkloadDependencies-Tests/WorkloadDependencies-Tests.csproj
           $(System.DefaultWorkingDirectory)/tests/Xamarin.Android.Tools.AndroidSdk-Tests/Xamarin.Android.Tools.AndroidSdk-Tests.csproj
         arguments: -c Debug $(XatBuildProperties)
